requirements.txt
cryptography
PyCryptodome